Renamed files and fixed applications path
