+++ examples/sycl/pvc/flash_attention_v2/pvc_flash_attn_runner.hpp
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2024 - 2024 Codeplay Software Ltd. All rights reserved.
+ * Copyright (c) 2024 - 2025 Codeplay Software Ltd. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,9 +32,9 @@
 
 #include "cutlass/epilogue/collective/default_epilogue.hpp"
 #include "cutlass/gemm/device/gemm_universal_adapter.h"
-#include "applications/pvc/flash_attention_v2/kernel/pvc_flash_attn_gemm_universal.hpp"
-#include "applications/pvc/flash_attention_v2/collective/pvc_flash_attn_epilogue.hpp"
-#include "applications/pvc/flash_attention_v2/collective/pvc_softmax_epilogue.hpp"
+#include "flash_attention_v2/kernel/xe_flash_attn_gemm.hpp"
+#include "flash_attention_v2/collective/xe_flash_attn_epilogue.hpp"
+#include "flash_attention_v2/collective/xe_flash_attn_softmax_epilogue.hpp"
 #include "cutlass/util/GPU_Clock.hpp"
 #include "cutlass/util/sycl_event_manager.hpp"
